Add files via upload
Determining the probability customers of a Telecom company are likely to churn

Models
Logistic Regression

This model was the starting point of our analysis and provided some useful initial indication on what the most important and most influential predictors were.
The logistic regression model was however found to be very prone to overfitting and required scaling the data to perform at its best.
Despite the very low computational power needed and the high speed of calculation, the results provided by the logistic regression model were fairly accurate and informative. Accuracy for the test sample was around 81.0%, while F1 score was around 57.7%. Due to significant variability, these values may nonetheless change between various iterations of the model.

SVM
The Support Vector Model was the most computationally intensive and slowest model out of those evaluated, requiring GPU acceleration to complete its modelling in a reasonable amount of time. The SVM also required scaling the data to work optimally and, given the many possible combinations of parameter values, necessitated GridSearch algorithms to identify the best possible combination of these to return the most accurate classification. Despite the complex and powerful modelling techniques, SVM does not provide substantially better results than other models and is therefore not chosen for further analysis. Accuracy for the test sample was 80.2% and F1 score was 59.8%.

Random Forest Model
The Random Forest model (and the Decision Tree model that lies at its core) strikes a good balance since it still allows substantial parameter tuning but is still relatively quick when constructing the model without requiring GPU acceleration. Random Forest also does not require scaling the data to work well, reducing the steps necessary to make the model perform efficiently. The accuracy of the predictions is nonetheless in line with other models. Accuracy for the test sample was 80.3% and F1 score was 61.3%. Most importantly, this model employs random sampling with replacement to reduce the variability of results by averaging many estimators, providing more consistent outputs and reducing the variance that affected previous models.